feat:connection campaign service with saga pattern

+++ Nistagram/campaign-service/src/main/java/app/campaignservice/service/ProfileServiceImpl.java
@@ -1,8 +1,13 @@
 package app.campaignservice.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import java.util.UUID;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import app.campaignservice.event.ProfileCanceledEvent;
 import app.campaignservice.dto.ProfileDTO;
 import app.campaignservice.model.Profile;
 import app.campaignservice.repository.ProfileRepository;
@@ -11,14 +16,17 @@ import app.campaignservice.repository.ProfileRepository;
 public class ProfileServiceImpl implements ProfileService{
 
 	private ProfileRepository profileRepository;
+	private final ApplicationEventPublisher publisher;
 	
 	@Autowired
-	public ProfileServiceImpl(ProfileRepository profileRepository) {
+	public ProfileServiceImpl(ProfileRepository profileRepository, ApplicationEventPublisher publisher) {
 		this.profileRepository = profileRepository;		
+		this.publisher = publisher;
 	}
 	
 	@Override
-	public void create(ProfileDTO profileDTO) {
+	@Transactional(rollbackFor = { Exception.class })
+	public void create(ProfileDTO profileDTO) throws Exception {
 		Profile profile = new Profile();
 		
 		profile.setUsername(profileDTO.username);
@@ -27,27 +35,33 @@ public class ProfileServiceImpl implements ProfileService{
 		profile.setBlocked(false);
 		
 		profileRepository.save(profile);
+		
+		publishProfileCanceled(profileDTO.username, "An error occurred while creating profile in campaign service.");
 	}
 	
+	private void publishProfileCanceled(String username, String reason) {
+		ProfileCanceledEvent event = new ProfileCanceledEvent(UUID.randomUUID().toString(), username,reason);     
+        publisher.publishEvent(event);
+    }
+	
 	@Override
+	@Transactional
 	public void updatePersonalData(String oldUsername, ProfileDTO profileDTO) {
-		Profile profile = profileRepository.findByUsername(oldUsername);
-				
+		Profile profile = profileRepository.findByUsername(oldUsername);		
 		profile.setUsername(profileDTO.username);
-		
 		profileRepository.save(profile);
 	}
 	
 	@Override
+	@Transactional
 	public void updateProfilePrivacy(ProfileDTO profileDTO) {
-		Profile profile = profileRepository.findByUsername(profileDTO.username);
-		
-		profile.setPublic(profileDTO.isPublic);
-		
+		Profile profile = profileRepository.findByUsername(profileDTO.username);	
+		profile.setPublic(profileDTO.isPublic);	
 		profileRepository.save(profile);
 	}
 	
 	@Override
+	@Transactional
 	public void blockProfile(String username) {
 		Profile profile = profileRepository.findByUsername(username);
 		profile.setBlocked(true);
@@ -71,9 +85,18 @@ public class ProfileServiceImpl implements ProfileService{
 	}
 
 	@Override
+	@Transactional
 	public void deleteByUsername(String username) {
 		Profile profile = profileRepository.findByUsername(username);
-		profileRepository.delete(profile);
-		
+		if(profile != null)
+			profileRepository.delete(profile);
+	}
+
+	@Override
+	@Transactional
+	public void createInfluencer(String username) {
+		Profile profile = profileRepository.findByUsername(username);	
+		profile.setInfluencer(true);	
+		profileRepository.save(profile);
 	}
 }
